backend/pyproject.toml
[tool.poetry]
name = "job-portal"
version = "1.0.0"
description = "A modern job posting website with FastAPI"
authors = ["Ajaysanepalle"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.104.1"
uvicorn = "0.24.0"
sqlalchemy = "2.0.23"
python-multipart = "0.0.6"
pyjwt = "2.8.1"
python-jose = "3.3.0"
passlib = "1.7.4"
bcrypt = "4.1.1"
python-dotenv = "1.0.0"
gunicorn = "21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
